pyproject.toml
[project]
name = "duckplexity"
version = "0.1.0"
description = ""
authors = [
    {name = "Codex",email = "codex@openai.com"}
]
readme = "README.md"

requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot (>=22.3,<23.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = true
packages = [
  { include = "bot" }
]

[project.scripts]
duckplexity = "bot.main:main"
